Extract API URL to .env
