--- .env
@@ -1,3 +1,0 @@
-REACT_APP_NETWORK_ID=56
-PUBLIC_URL=''
-IMAGE_INLINE_SIZE_LIMIT=0
