+++ .env.example
@@ -1,0 +1,4 @@
+REACT_APP_NETWORK_ID=56
+REACT_APP_API_URL=https://api.beefy.finance
+PUBLIC_URL=''
+IMAGE_INLINE_SIZE_LIMIT=0
